Finally sim-engine tests works
