+++ src/test/scala/download/TestDownloadCSV.scala
@@ -1,0 +1,5 @@
+package download
+
+import org.scalatest.funspec.AnyFunSpec
+
+class TestDownloadCSV extends AnyFunSpec
